combined unit tests, and formatted
